module vente mobile layout

+++ sales-android/src/main/java/com/kobe/warehouse/sales/data/model/ClientTiersPayant.kt
@@ -15,7 +15,7 @@ data class ClientTiersPayant(
     val id: Long? = null,
 
     @SerializedName("customerId")
-    val customerId: Long? = null,
+    val customerId: Int? = null,
 
     @SerializedName("tiersPayantId")
     val tiersPayantId: Long,
